Bug fixing

+++ src/Acme/FindMartialBundle/Controller/ArtController.php
@@ -200,6 +200,9 @@ class ArtController extends Controller
      */
     public function deleteAction(Request $request, $id)
     {
+
+        $securityContext = $this->get('security.context');
+        
         $form = $this->createDeleteForm($id);
         $form->bind($request);
 
@@ -209,6 +212,10 @@ class ArtController extends Controller
 
             if (!$entity) {
                 throw $this->createNotFoundException('Unable to find Art entity.');
+            }
+
+            if (false === $securityContext->isGranted('DELETE', $entity)) {
+                throw new AccessDeniedException();
             }
 
             $em->remove($entity);
