+++ src/Acme/FindMartialBundle/Controller/ClubController.php
@@ -233,6 +233,8 @@ class ClubController extends Controller
      */
     public function deleteAction(Request $request, $id)
     {
+        $securityContext = $this->get('security.context');
+
         $form = $this->createDeleteForm($id);
         $form->bind($request);
 
@@ -243,6 +245,11 @@ class ClubController extends Controller
             if (!$entity) {
                 throw $this->createNotFoundException('Unable to find Club entity.');
             }
+
+            if (false === $securityContext->isGranted('DELETE', $entity)) {
+                throw new AccessDeniedException();
+            }
+
 
             $em->remove($entity);
             $em->flush();
